requirements.txt
dash==0.39.0
plotly
gunicorn
numpy
scipy
pandas
mysql-connector